+++ config/routes.js
@@ -1,6 +1,6 @@
 // eslint-disable-next-line new-cap
 const router = require('express').Router();
-const gpMiddleware = require('../app/middleware/gp-practice');
+const gpMiddleware = require('../app/middleware/gp');
 
 router.get('/gpdetails/:gpId',
   gpMiddleware.getUrl,
